refactor(models, dao): Restructure and relocate `BranchDao` for better modularity
- Moved `BranchDao` to `models.dao` package for alignment with package organization.
- Updated references in `SuperAdminController` to reflect the new `BranchDao` path.
- Adjusted `module-info.java` to export and open the new package structure.
