+++ src/main/java/module-info.java
@@ -11,8 +11,10 @@ module mx.edu.utez.warehousemanagerfx {
 
     opens mx.edu.utez.warehousemanagerfx to javafx.fxml;
     exports mx.edu.utez.warehousemanagerfx;
-    opens mx.edu.utez.warehousemanagerfx.models to javafx.fxml;
-    exports mx.edu.utez.warehousemanagerfx.models;
     opens mx.edu.utez.warehousemanagerfx.controllers to javafx.fxml;
     exports mx.edu.utez.warehousemanagerfx.controllers;
+    opens mx.edu.utez.warehousemanagerfx.models to javafx.fxml;
+    exports mx.edu.utez.warehousemanagerfx.models;
+    exports mx.edu.utez.warehousemanagerfx.models.dao;
+    opens mx.edu.utez.warehousemanagerfx.models.dao to javafx.fxml;
 }
