+++ src/main/java/mx/edu/utez/warehousemanagerfx/Models/dao/BranchDao.java
@@ -1,5 +1,6 @@
-package mx.edu.utez.warehousemanagerfx.models;
+package mx.edu.utez.warehousemanagerfx.models.dao;
 
+import mx.edu.utez.warehousemanagerfx.models.Branch;
 import mx.edu.utez.warehousemanagerfx.utils.database.OracleDatabaseConnectionManager;
 
 import java.sql.Connection;
